Add data loader for the training data

+++ utils.py
@@ -16,14 +16,18 @@ def get_data_loader(data_path, batch_size, tokenizer, shuffle=True, max_len=20):
     """
     data = np.loadtxt(data_path, delimiter='\t', dtype=str)
     X, y = data[:, -1], data[:, 1]
+   
     X = tokenizer.batch_encode_plus(
         X.tolist(), max_length=max_len, truncation=True, padding='max_length')
     X, mask = X['input_ids'], X['attention_mask']
     X = torch.tensor(np.array(X))
     mask = torch.tensor(np.array(mask))
-    y = torch.tensor(np.array(y, dtype=int)).long()
-    # y = torch.tensor(np.array(y, dtype=int))
+    y = torch.tensor(np.array(y, dtype=int))
     data = torch.utils.data.TensorDataset(X, mask, y)
+    
     data_loader = torch.utils.data.DataLoader(
         data, batch_size=batch_size, shuffle=shuffle)
+    
     return data_loader
+    
+
